requirements.txt
scipy==1.15.2
numpy==1.24.0 
matplotlib==3.10.0
pillow==10.4.0
opencv-python==4.11.0.86
tqdm==4.67.1
zerorpc==0.6.3
wandb==0.19.9
openai==1.71.0
google-genai==1.11.0
gevent==24.11.1
hydra-core==1.3.2
imageio[ffmpeg]==2.37.0
omegaconf==2.3.0
deepdiff==8.4.2